Unify verify_ticket_id() method for all ticketing tools
